requirements.txt
python-telegram-bot==21.7
aiosqlite==0.20.0
python-dotenv==1.0.1
aiohttp==3.9.1
